+++ tests/testthat/test-get_sample_size.R
@@ -29,22 +29,25 @@ test_that("get_sample_size works", {
     #### compute_n: giant ####
     dat7 <- echodata::get_sample_size(dat = dat,
                                       compute_n = "giant",
-                                      return_only = max)
+                                      return_only = max, 
+                                      na.rm=TRUE)
     testthat::expect_equal(dat7, 108310)
     #### compute_n: metal ####
     dat8 <- echodata::get_sample_size(dat = dat,
                                       compute_n = "metal",
-                                      return_only = max)
+                                      return_only = max,
+                                      na.rm=TRUE)
     testthat::expect_equal(dat8, 216621)
     #### compute_n: sum ####
     dat9 <- echodata::get_sample_size(dat = dat,
                                       compute_n = "sum",
-                                      return_only = max)
+                                      return_only = max,
+                                      na.rm=TRUE)
     testthat::expect_equal(dat9, 1474097)
     #### compute_n: number ####
     dat10 <- echodata:: get_sample_size(dat = dat,
                                       compute_n = 10000,
-                                      return_only = max)
+                                      return_only = max,
+                                      na.rm=TRUE)
     testthat::expect_equal(dat9, 1474097)
-    
 })